refactor: move the internal greeter package up one level
